Use papyros structure
