+++ src/App.ts
@@ -59,7 +59,10 @@ async function startPapyros(): Promise<void> {
         standAloneOptions: {
             parentElementId: "root"
         },
-        darkMode: darkMode
+        darkMode: darkMode,
+        traceOptions: {
+            parentElementId: "trace-root"
+        }
     });
     setUpEditor(papyros.codeRunner.editor, LOCAL_STORAGE_KEYS.code);
     papyros.codeRunner.editor.focus();
