+++ src/CodeRunner.ts
@@ -26,6 +26,7 @@ import {
     renderWithOptions
 } from "./util/Rendering";
 import {OutputManager} from "./OutputManager";
+import {TraceViewer} from "./TraceViewer";
 
 interface DynamicButton {
     id: string;
@@ -50,6 +51,7 @@ interface CodeRunnerRenderOptions {
      * RenderOptions for the output field
      */
     outputOptions: RenderOptions;
+    traceOptions: RenderOptions;
 }
 
 /**
@@ -97,6 +99,7 @@ export class CodeRunner extends Renderable<CodeRunnerRenderOptions> {
      * Component to handle output generated by the user's code
      */
     public readonly outputManager: OutputManager;
+    public readonly traceViewer: TraceViewer;
     /**
      * The backend that executes the code asynchronously
      */
@@ -175,6 +178,7 @@ export class CodeRunner extends Renderable<CodeRunnerRenderOptions> {
         this.previousState = RunState.Ready;
         this.runStartTime = new Date().getTime();
         this.state = RunState.Ready;
+        this.traceViewer = new TraceViewer();
     }
 
     /**
@@ -364,6 +368,7 @@ export class CodeRunner extends Renderable<CodeRunnerRenderOptions> {
         this.editor.setPanel(rendered);
         // Set language again to update the placeholder
         this.editor.setProgrammingLanguage(this.programmingLanguage);
+        this.traceViewer.render(options.traceOptions);
         return rendered;
     }
